% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiho_cluster_analysis.R
\name{get_pca_components}
\alias{get_pca_components}
\title{get_pca_components}
\usage{
get_pca_components(df)
}
\arguments{
\item{df}{a data frame with n rows (data points) and p columns
(numeric variables)

OUTPUT}
}
\value{
a data frame with pca coordinates
}
\description{
DESCRIPTION
this function calculates and returns the pca components from
the input data frame df. This is helpful to visualize the outcome
of the clustering trask performed by the KmeansElbow function
}
\details{
PARAMETERS
}
