% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiho_cluster_analysis.R
\name{KmeansElbow}
\alias{KmeansElbow}
\title{KmeansElbow}
\usage{
KmeansElbow(
  x,
  rep = 10,
  max.clusters = 20,
  nstarts = 10,
  scaling.method = "na",
  decision.method = "knee",
  plot.pca = FALSE
)
}
\arguments{
\item{x:}{numeric matrix of data, or an object that can be
coerced to such a matrix (such as a numeric vector or a data
frame with all numeric columns).}

\item{rep:}{maximum number of iterations (default 10;
iter.max in kmeans()).
After selecting the initial k clusters, the model performs
an iterative process of computing and selecting the mean
of each cluster as the new centroid for the next interation.
This parameter controls how many times such iteration
should be performed.}

\item{max.clusters:}{maximum number of clusters to evaluate.
Default 20.}

\item{nstarts:}{how many random sets should be chosen?
Default 10.
For example, when finding k clusters, the model initially
randomly selects k data points from x to conform the inital
clusters. This parameter specifies how many times the model
should perform this random selection. It is recommended to
use nstarts > 1. This is different to the parameter rep (see above).}

\item{scaling.method:}{feature scaling method. In case the data
has not been scaled, whether to use "standardization" or
"z-score". By default, the function assumes that the data has
already been prepared for the Kmeans model (i.e, "na").}

\item{decision.method:}{whether "silhouette" or "knee" (default)
method. Knee method is based on the implementation by Coffey et al
(2019).}

\item{plot.pca:}{whether to perform a pca plot colored according
to the k clusters (default FALSE)}

\item{elbow.to.pdf:}{boolean, save elbow figure to pdf (default FALSE)}

\item{filename:}{name of the pdf where to save the elbow figure.
Default "elbow"}

\item{outdir:}{output directory where to store the elbow figure.
Default "." (i.e., current directory)}

}
\value{
kmeans object (see kmeans function from stats package)

COMMENTS ON THE K-MEANS ALGORITHM
The algorithm of Hartigan and Wong (1979) is used by default
Kmeans is sensitive to outlayers and to initialization of the
centroids
Kmeans is scale dependent and isnt suitable for data of varying
shapes and densities
Different solutions can converge with random initialization
metrics:
inertia: sum of squared errors for each cluster (within cluster sse)
Silhouette coefficient: measure of how far apart clusters are
high silhouette score means that clusters are well separated
}
\description{
DESCRIPTION
This function performs unsupervised clustering
using the K-means model. The optimal number of
clusters is selected automatically according to
two alternative methods decided by the user:
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    1) Silhouette score (default method)
    2) Knee method (as in DeepSqueak)
}\if{html}{\out{</div>}}

ARGUMENTS
}
\examples{
Assuming that x is a data frame with numeric variables only
km <- KmeansElbow(x) # run the model
str(km) # check data structure
x <- cbind(km$cluster, x) # add clustering results (IDs) to x
head(x)

}
